izaj-desktop/backend/python/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
supabase==1.0.4
python-dotenv>=1.0.0
pandas>=2.2.0
numpy>=1.26.0
pydantic>=2.5.0
python-multipart>=0.0.6
httpx>=0.24.0,<0.25.0
